+++ src/Psl/Unix/Socket.php
@@ -74,12 +74,9 @@ final class Socket
      *
      * This consumes the socket; it cannot be reused after calling listen().
      *
-     * @param int<1, max> $backlog Maximum length of the queue of pending connections.
-     * @param int<1, max> $idleConnections Maximum number of idle connections to buffer.
-     *
      * @throws Network\Exception\RuntimeException If listen fails.
      */
-    public function listen(int $backlog = 128, int $idleConnections = 256): ListenerInterface
+    public function listen(ListenConfiguration $configuration = new ListenConfiguration()): ListenerInterface
     {
         $this->ensureNotConsumed();
         $this->consumed = true;
@@ -91,12 +88,12 @@ final class Socket
         }
 
         $context = ['socket' => [
-            'backlog' => $backlog,
+            'backlog' => $configuration->backlog,
         ]];
 
         $stream = Network\Internal\server_listen("unix://{$this->bindPath}", $context);
 
-        return new Internal\Listener($stream, $idleConnections);
+        return new Internal\Listener($stream, $configuration->idleConnections);
     }
 
     /**